Finish remove, rename file. Create share part

+++ app/Http/Controllers/FileController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
 
-use App\File;
+use App\Files;
 use App\Dir;
 
 class FileController extends Controller
@@ -21,14 +23,24 @@ class FileController extends Controller
 
             $file = $request->file('fileUpload');
 
-            $fileName = $file->getClientOriginalName();
+            if ($request->docsName != '') {
+                $fileName = $request->docsName;
+            } else {
+                $fileName = $file->getClientOriginalName();
+            }
+
+            
             $fileExtension = $file->getClientOriginalExtension();
 
             // Tạo tên file
             $pages = range(1,20);
             shuffle($pages);
             $prefix = array_shift($pages);
-            $fileNameToSave = time() . Auth::user()->id . $prefix . '_' . $fileName;
+            if ($request->docsName != '') {
+                $fileNameToSave = time() . Auth::user()->id . $prefix . '_' . $request->docsName;
+            } else {
+                $fileNameToSave = time() . Auth::user()->id . $prefix . '_' . $fileName;
+            }
 
             $file->move('fileUploaded', $fileNameToSave);
 
@@ -43,7 +55,7 @@ class FileController extends Controller
             }
 
             // Lưu thông tin file vào File database
-            $fileDb = new File;
+            $fileDb = new Files;
             $fileDb->name = $fileNameToSave;
             $fileDb->fileName = $fileName;
             $fileDb->dir = $dir;
@@ -105,7 +117,7 @@ class FileController extends Controller
             $dir = $request->currentDir;
 
             // Lưu thông tin file vào File database
-            $fileDb = new File;
+            $fileDb = new Files;
             $fileDb->name = $fileNameToSave;
             $fileDb->fileName = $fileName;
             $fileDb->dir = $dir;
@@ -150,12 +162,73 @@ class FileController extends Controller
 
     public function removeType(Request $request) {
         $dir = new Dir;
-        $dirToRemove = $dir->where('dir', $request->dirToRemove)->delete();
+        $dirToRemove = $dir->where('owner', Auth::user()->name)->where('dir', $request->dirToRemove)->delete();
 
         return redirect('library');
     }
 
     public function typeData($type) {
         return view('user-function/library', ['title' => 'Thư viện', 'typeDir' => $type]);
+    }
+
+    public function removeItem(Request $request) {
+        $fileDb = new Files;
+
+        $fileDir = $request->itemDirSelected;
+        $fileName = $request->itemNameSelected;
+        
+        $fileToRemove = $fileDb->where('owner', Auth::user()->name)->where('name', $fileName)->where('dir', $fileDir)->delete();
+        
+        $fileDirToRemove = "/fileUpload/" . $fileName;
+        File::delete($fileDirToRemove);
+        return redirect()->back();
+    }
+
+    public function renameItem(Request $request) {
+        $this->validate($request, [
+            'name' => 'required|min:3'
+        ], [
+            'name.required' => 'Không nhận được dữ liệu',
+            'name.min'      => 'Vui lòng nhập tối thiểu 3 kí tự'
+        ]);
+
+        $fileDb = new Files;
+
+        // Xử lí dữ liệu
+        $fileDir = $request->itemDirSelected;
+        $fileName = $request->itemNameSelected;
+        $newFileName = $request->name;
+
+
+        $fileToRename = $fileDb->where('owner', Auth::user()->name)->where('name', $fileName)->where('dir', $fileDir)->get();
+        $id = $fileToRename[0]->id;
+
+        $fileToRename = Files::find($id);
+
+        $fileRealName = $fileToRename->name;
+        $oldFileName = $fileToRename->fileName;
+        $oldFileRealName = $fileToRename->name;
+
+        // Dữ liệu để rename
+        $newFileName .= '.' . $fileToRename->type;
+        $newFileRealName = str_replace($oldFileName, $newFileName, $fileRealName);
+
+        echo $newFileName . "<br/>";
+        echo $newFileRealName . "<br/>";
+        // dd($fileToRename);
+
+        // $fileToRename->update(['fileName' => $newFileName, 'name' => $newFileRealName]);
+        $fileToRename->fileName = $newFileName;
+        $fileToRename->name = $newFileRealName;
+        $fileToRename->save();
+
+        echo "<hr/>";
+
+        echo $fileToRename->fileName . "<br/>";
+        echo $fileToRename->name . "<br/>";
+
+        rename(public_path("fileUploaded/$oldFileRealName"), public_path("fileUploaded/$newFileRealName"));
+        
+        return redirect()->back();   
     }
 }
